move ad hoc alignments to adjustments.py
